+++ JSM.Insight/Public/New-ObjectType.ps1
@@ -47,7 +47,7 @@ function New-ObjectType {
     
     end {
         try {
-            $response = Invoke-RestMethod -Uri $Request.Uri -Headers $headers -Method POST
+            $response = Invoke-RestMethod -Uri $Request.Uri -body $RequestBody -Headers $headers -Method POST
         }
         catch {
             Write-Verbose "[$($MyInvocation.MyCommand.Name)] Failed"